common::Vecをテンプレートクラス化 common::Vec is made into a template class

+++ src/Collision/Collision.cpp
@@ -35,7 +35,7 @@ bool collision::isInCollision(collision::Rectangle* rectangle1, collision::Recta
     return true;
 }
 
-float collision::isInCollision(collision::Rectangle* rectangle, collision::Circle* circle){
+double collision::isInCollision(collision::Rectangle* rectangle, collision::Circle* circle){
     // 円の右端の画面x座標
     double circle_right_display_x = circle->center_display_x + circle->radius;
     // 円の上端の画面y座標
@@ -58,7 +58,7 @@ float collision::isInCollision(collision::Rectangle* rectangle, collision::Circl
             // そうでない、すなわち矩形の角と円形が衝突している場合
             }else{
                 // 返り値
-                float dir;
+                double dir;
                 // 右上の角
                 if(circle->center_display_x > rectangle->right_display_x && circle->center_display_y < rectangle->top_display_y){
                     dir = atan2(circle->center_display_y - rectangle->top_display_y, circle->center_display_x - rectangle->right_display_x);
@@ -103,11 +103,11 @@ float collision::isInCollision(collision::Rectangle* rectangle, collision::Circl
     }
 }
 
-float collision::isInCollision(collision::Circle* circle, collision::Rectangle* rectangle){
+double collision::isInCollision(collision::Circle* circle, collision::Rectangle* rectangle){
     return isInCollision(rectangle, circle);
 }
 
-float collision::isInCollision(collision::Circle* circle1, collision::Circle* circle2){
+double collision::isInCollision(collision::Circle* circle1, collision::Circle* circle2){
     // 2乗したまま比較
     // 半径の合計
     uint32_t radius_sum = (circle1->radius + circle2->radius) * (circle1->radius + circle2->radius);
@@ -116,7 +116,7 @@ float collision::isInCollision(collision::Circle* circle1, collision::Circle* ci
     // 衝突判定
     if(dist <= radius_sum){
         // 衝突する場合力積の方向を計算
-        float dir = atan2f(circle1->center_display_y - circle2->center_display_y, circle1->center_display_x - circle2->center_display_x);
+        double dir = atan2f(circle1->center_display_y - circle2->center_display_y, circle1->center_display_x - circle2->center_display_x);
         if(dir < 0){
             return dir + M_PI * 2;
         }else{
